+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.29] - 2026-01-11
+
+### Fixed
+
+- YouTubeダウンロード時のクッキー取得をプラットフォーム対応
+  - macOS: Safari
+  - Windows: Edge
+  - Linux: Firefox
+
+---
+
 ## [2.1.28] - 2026-01-11
 
 ### Added